Exercise4.21 Done, exercise 4 Done

+++ exercise4/blogilista/controllers/blogs.js
@@ -39,15 +39,27 @@ blogsRouter.post('', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-  const id = request.params.id
+  try {
+    const id = request.params.id
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+    const user = await User.findById(decodedToken.id)
+    const bleg = await Blog.findById(id)
+  } catch (e) {
+    response.status(401).json('You can only delete your own blogs!')
+  }
+  if (decodedToken.id === bleg.user.toString()) {
+    const result = await Blog.findByIdAndRemove(id, (err) => {
+      if (err) {
+        return response.status(404).json('Not found')
+      } else {
+        return response.status(200).json('Successfully deleted')
+      }
+    })
+  } else {
+    response.status(401).json('You can only delete your own blogs!')
+  }
 
-  const result = await Blog.findByIdAndRemove(id, (err) => {
-    if (err) {
-      return response.status(404).json('Not found')
-    } else {
-      return response.status(200).json('Successfully deleted')
-    }
-  })
+
 })
 
 blogsRouter.put('/:id', async (request, response) => {
